requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.1
redis==5.0.1
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
celery==5.3.4
flower==2.0.1
pandas==2.1.4
beautifulsoup4==4.12.2
validators==0.22.0
xlsxwriter==3.1.9
psutil==5.9.6
plotly==5.17.0
websockets==12.0
